Remove length from the header struct.

+++ src/v2/error.rs
@@ -15,11 +15,11 @@ pub enum ParseError {
     #[error("Invalid protocol {0:X}. Protocol must be one of: Unspecified, Stream, or Datagram.")]
     Protocol(u8),
     #[error("Header does not contain the advertised length of the address information and TLVs (has {1} out of {0} bytes).")]
-    Partial(u16, usize),
+    Partial(usize, usize),
     #[error(
         "Header length of {0} bytes cannot store the {1} bytes required for the address family."
     )]
-    InvalidAddresses(u16, usize),
+    InvalidAddresses(usize, usize),
     #[error("Header is not long enough to contain enough TLV {0} with lengh {1}.")]
     InvalidTLV(u8, u16),
     #[error("Header contains leftover {0} bytes not accounted for by the address family or TLVs.")]
